refactor: API 비즈니스 로직 분리
